Move k/test/integration/federation to k/federation/test/integration
