+++ federation/test/integration/main_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package federation
+package integration
 
 import (
 	"testing"
